+++ R/example_setup.R
@@ -113,7 +113,7 @@ example_setup.pf_smoother_two_filter <- function(.fun, .connect = TRUE) {
   #### Simulate movements
   state      <- "StateXY"
   model_move <-
-    move_xy(.mobility = "750.0",
+    model_move_xy(.mobility = "750.0",
             .dbn_length = "truncated(Gamma(1, 250.0), upper = 750.0)",
             .dbn_heading = "Uniform(-pi, pi)")
   sim_path_walk(.map = map,