+++ dev/001-benchmark.R
@@ -72,7 +72,7 @@ moorings <- run(file = here_data_raw("benchmark", "moorings.rds"),
 state      <- "StateXY"
 mobility   <- 750
 model_move <-
-  move_xy(.mobility     = mobility,
+  model_move_xy(.mobility     = mobility,
           .dbn_length   = glue("truncated(Gamma(15.0, 250.0), upper = {mobility})"),
           .dbn_heading  = "Uniform(-pi, pi)")
 # Simulate path