+++ man/examples/example-ModelMove.R
@@ -21,51 +21,51 @@ if (patter_run(.geospatial = FALSE)) {
   # > This is required as a component of all movement models
   set_map(map)
 
-  #### Example (1a): Use `move_xy()` with default options
-  # `move_*()` functions simply return a character string of Julia code
+  #### Example (1a): Use `model_move_xy()` with default options
+  # `model_move_*()` functions simply return a character string of Julia code
   # (Downstream functions can evaluate this code, as shown below)
-  move_xy()
+  model_move_xy()
 
-  #### Example (1b): Customise `move_xy()`
+  #### Example (1b): Customise `model_move_xy()`
   # Use a truncated normal distribution for step lengths:
-  move_xy(.mobility = "750.0",
+  model_move_xy(.mobility = "750.0",
           .dbn_length = "truncated(Normal(250, 50), lower = 0.0, upper = 750.0)")
   # Use an exponential distribution for step lengths
-  move_xy(.mobility = "750.0",
+  model_move_xy(.mobility = "750.0",
           .dbn_length = "truncated(Exponential(0.01), upper = 750.0)")
   # Use a biased random walk
-  move_xy(.dbn_heading = "VonMises(0, 1)")
+  model_move_xy(.dbn_heading = "VonMises(0, 1)")
   # Get help on a distribution in Julia:
   julia_help("Exponential")
 
-  #### Example (2): Use `move_xyz()`
+  #### Example (2): Use `model_move_xyz()`
   # Use default options
-  move_xyz()
+  model_move_xyz()
   # Customise model components
-  move_xyz(.mobility = "750.0",
+  model_move_xyz(.mobility = "750.0",
            .dbn_length = "truncated(Exponential(0.01), upper = 750.0)",
            .dbn_z = "truncated(Gamma(1.0, 250.0), lower = 0.0, upper = 350.0)")
 
-  #### Example (3): Use `move_cxy()`
+  #### Example (3): Use `model_move_cxy()`
   # Use default options
-  move_cxy()
+  model_move_cxy()
   # Customise model components
-  move_cxy(.mobility = "750.0",
+  model_move_cxy(.mobility = "750.0",
             .dbn_length = "truncated(Normal(250, 50), lower = 0.0, upper = 750.0)",
             .dbn_heading_delta = "Normal(0, 0.25)")
 
-  #### Example (4): Use `move_cxyz()`
+  #### Example (4): Use `model_move_cxyz()`
   # Use default options
-  move_cxyz()
+  model_move_cxyz()
   # Customise model components
-  move_cxyz(.mobility = "750.0",
+  model_move_cxyz(.mobility = "750.0",
             .dbn_length = "truncated(Normal(250, 50), lower = 0.0, upper = 750.0)",
             .dbn_heading_delta = "Normal(0, 0.25)",
             .dbn_z_delta = "Normal(0, 2.5)")
 
   #### Example (4): Visualise movement model component distributions
   # See `?plot.ModelMove`
-  plot(move_xy())
+  plot(model_move_xy())
 
   #### Example (5): Visualise movement model realisations (trajectories)
   if (patter_run()) {
@@ -87,9 +87,9 @@ if (patter_run(.geospatial = FALSE)) {
                  .n_path = 2L, .one_page = FALSE)
     # Compare different movement models via `sim_path_walk()`
     pp <- par(mfrow = c(2, 2))
-    args$.model_move <- move_xy()
+    args$.model_move <- model_move_xy()
     do.call(sim_path_walk, args)
-    args$.model_move <- move_xy(.dbn_heading = "VonMises(0.1, 0.1)")
+    args$.model_move <- model_move_xy(.dbn_heading = "VonMises(0.1, 0.1)")
     do.call(sim_path_walk, args)
     par(pp)